microsoft-fabric/integracio_per_consultar_dades/exemples/sqlserver/python/requirements.txt
pyodbc
azure-identity
msal